Use getminercandidate rather than getblocktemplate to get next block diff
getminercandidate requires less locking (it uses a more granular locking
scheme), whereas getblocktemplate use cs_main for the entirety of the time
It also provides the entire block to the explorer, where it just the difficulty
for the next block (in bits encoding) would be needed, which is what getminercandidate.

Another point worth underlining is that since the call to getminercandidate
is not going to proivde "mission critical" data to the explorer but just a glimpse
of what difficulty nexa DAA (ASERT) computed for the next block, we should also
set mining.unsafeGetBlockTemplate to true.

This tweak default value is set to false and this is due to the fact that
we don't want a miner to mine on top of a stale chaintip.

This of course comes at a price, in fact when set to false it means we
need further locking in getminingcandidate to make sure we are well
connected to other nodes. Since the explorer is not a miner we could set
a different balance here and lean a bit more on the side of performance
rather than correctness.