Fix JSON parsing/stringify issue with bigint fields
The problem we are trying to solve is related to the fact that
the JSON standard does not define/specify integer precision.
In javascript there are 2 types to represent numbers: `Number` and
`BigInt`.

The former is a 64-bit floating point value which can contains
up to 16 digit, the latter is the same as `long` in C++ or Java.
It lets you store up to 20 digits.

Problem is that JSON.{parse,stringify} use `Number` when parsing
the data coming back from JSON RPC requests.

This is an example of potential loss of information when parsing
JSON RPC data received form nexad:

```
JSON.stringify(99999999989999831)
'99999999989999820'
```

This would lead to a problem for all those cases where we are dealing
with numerical field the exceed the js `Number` type domain range.

This happen quite more frequently than before since we started to
see more token transactions in the Nexa network. In particular this
happen when people mint coins for a new token class, e.g. this is the
txidem of such a transaction:

e92cff98f2395cde428038bcebf22f68905dfee2939120235982f9aeaee842f6

And this is the "scriptPubKey" field for the second output of the above
txn as displayed by the explorer before the changes included in this
commit:

```
  "scriptPubKey": {
      "asm": "ee7cf7b94ba9120e8ffedfa2a813f3b48a3018cb3bc24b6358694df1abc30000 99999999989999831 1 0e470fe70f1ea4eea5730ce3a3dd8a8b8e6e9c8f",
      "hex": "20ee7cf7b94ba9120e8ffedfa2a813f3b48a3018cb3bc24b6358694df1abc3000008d768f15c7845630151140e470fe70f1ea4eea5730ce3a3dd8a8b8e6e9c8f",
      "type": "scripttemplate",
      "scriptHash": "pay2pubkeytemplate",
      "argsHash": "0E470FE70F1EA4EEA5730CE3A3DD8A8B8E6E9C8F",
      "group": "nexa:trh8eaaefw53yr50lm0692qn7w6g5vqcevauyjmrtp55mudtcvqqqmd062t97",
      "groupQuantity": 99999999989999820,
      "groupAuthority": 0,
      "addresses": [
          "nexa:nqtsq5g5perslec0r6jwaftnpn368hv23w8xa8y0xadr7ysn"
      ]
  }
```

This is the same output gathered from `nexa-cli getrawtransaction`
command

```
      "scriptPubKey": {
        "asm": "ee7cf7b94ba9120e8ffedfa2a813f3b48a3018cb3bc24b6358694df1abc30000 99999999989999831 1 0e470fe70f1ea4eea5730ce3a3dd8a8b8e6e9c8f",
        "hex": "20ee7cf7b94ba9120e8ffedfa2a813f3b48a3018cb3bc24b6358694df1abc3000008d768f15c7845630151140e470fe70f1ea4eea5730ce3a3dd8a8b8e6e9c8f",
        "type": "scripttemplate",
        "scriptHash": "pay2pubkeytemplate",
        "argsHash": "0E470FE70F1EA4EEA5730CE3A3DD8A8B8E6E9C8F",
        "group": "nexa:trh8eaaefw53yr50lm0692qn7w6g5vqcevauyjmrtp55mudtcvqqqmd062t97",
        "groupQuantity": 99999999989999831,
        "groupAuthority": 0,
        "addresses": [
          "nexa:nqtsq5g5perslec0r6jwaftnpn368hv23w8xa8y0xadr7ysn"
        ]
      },
```

As you can see there's a loss of information for the `groupQuantity` field.

To solve the above problem we leverage an experimental feature
that is going to be included in the ECMAScript standard soon. Currently
it is included in nodejs ver >=20. The feature is not enabled by default
hence node need to be executed using the --harmony-json-parse-with-source flag.

This feature make it possible for the JSON `reviver` and `replacers`
functions to access to the original (pristine) JSON text before that
any serialization/deserialization operation is executed on the text.

Other solutions have been considered, namely the use of `lossless-json`,
but this would have meant the need to maintain a fork of `jayson` package
modified so that it used the aforementioned package.

We thought that since `jayson` let you pass your own version of
the `reviver` and `replacers` functions as options for its http
connection module, it is a more elegant solution to do just that
rather than adding one additional dep to the project (lossless-json)
and i the burden to maintain a forked version of a package.

We needed also to use another js package to access Redis `MessagePack`
tools. The new one is `msgpackr` and it has been adopted in place of
`msgpack-lite` because it does not support the BigInt type.