Make NEXA token group ID selectable
For space layout reasons the cashaddr encoding og a group id is
always shown in a compact form in the transaction view, e.g.

nexa:tp6j6...qqq26qrjtx

rathe than:

nexa:tp6j63e9nvkkqmukw8hg8kt95zcm8aq0mwj645qvqmp07p7zssqqqq26qrjtx

Previous to this change, the latter form was never showed and users
need to go on the json tab and search through the vout array to fetch
it. For the group id related to transaction inputs this was not even
possible.

This commit introduce a change in the UI so that by clicking onf the
span that shows the group id in compact form a persistent popover appear
that contains both the hex and cashaddr encode of the group ID, and
both of them are selectable.