Make the explorer working on testnet again
Since MR !89 we broke the explorer when working on testnet. The root
cause was there even before those set of changes, but MR !89 had the
unfortunate result to expose the problem.

Basically we decide which transport schema to use to connect to the
electrum server based on an hard-coded value for the default connection
porti for the TCP scheme.

This was incorrect cause while it is true that connection port changes
based on the transport scheme (see https://bitcoinunlimited.gitlab.io/rostrum/ports/),
it also changes based on the chain nexad is working on (eg nexa,
testnet, regtest).

Hence the code has been changed so that it chooses the transport scheme based on the
result of the parsing of the server URI specified in the configuration file.

https://electrum-cash.gitlab.io/network/global.html#TransportScheme